OOBs adjusted to proper AI order.
So German Panzer divisions start with 8xLt (Pz I) and 3xMotInf.  They add a Medium Battalion with IV-A, then add a 4th infantry battalion with IV-D. When Bison is researched, they add a SPArt Bn. When PzJg I is researched, they get a LtTd Bn.

By now... AGs are skipped for AI.